Moved tests + various small changes
Moved tests to better positions
Changed dice back
Adjusted UI
