+++ app/src/main/java/com/ss20/se2/monopoly/models/Dice.java
@@ -1,12 +1,8 @@
 package com.ss20.se2.monopoly.models;
 
-import java.util.Random;
-
 public class Dice{
 
 	public int roll(){
-		Random r = new Random();
-		int diceroll = r.nextInt(5) + 1;
-		return diceroll;
+		return (int) (6 * Math.random()) + 1;
 	}
 }
