+++ app/src/test/java/com/ss20/se2/monopoly/GameStateUnitTest.java
@@ -167,4 +167,35 @@ public class GameStateUnitTest{
 		assertEquals(2, GameState.getInstance().countOfPlayersRailroads(player));
 
 	}
+
+	@Test
+	public void GameboardTest(){
+		Gameboard gameboard = GameState.getInstance().getGameboard();
+		gameboard.gameboardArray[0] = new GamePiece("Player 1");
+
+		gameboard.getPosition("Player 1");
+		assertEquals(0, gameboard.getPosition("Player 1"));
+		gameboard.move("Player 1", 5);
+		assertEquals(5, gameboard.getPosition("Player 1"));
+		gameboard.move("Player 1", 20);
+		assertEquals(25, gameboard.getPosition("Player 1"));
+		gameboard.move("Player 1", 16);
+		assertEquals(1, gameboard.getPosition("Player 1"));
+
+		List<Street> streets = new ArrayList<>();
+		List<Railroad> railroads = new ArrayList<>();
+		List<Utility> utilities = new ArrayList<>();
+
+		Utility u1 = new Utility("Wasserwerk", 200, 100, false, UtilityType.WATER_WORKS);
+		Railroad r1 = new Railroad("Hauptbahnhof", 200, 100, false);
+		Street s1 = new Street("Esterhazy", 100, 50, false,175, 300, "green");
+
+		streets.add(s1);
+		assertEquals(s1, gameboard.getStreets().get(0));
+		railroads.add(r1);
+		assertEquals(r1, gameboard.getRailroads().get(0));
+		utilities.add(u1);
+		assertEquals(u1, gameboard.getUtilities().get(0));
+
+	}
 }
